pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "doorbell-notifier"
version = "2.0.0"
description = "MQTT doorbell desktop notifier"
requires-python = ">=3.9"
dependencies = ["paho-mqtt>=2.0", "PyYAML>=6.0", "emoji>=2.0"]

[project.scripts]
doorbell-notifier = "doorbell_notifier.main:main"
